refactor: extract render-thread feature runtime
