+++ internal/src/features/runtime/FeatureRuntime.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace FeatureRuntime {
+
+void ApplyOverrides();
+bool PollSocketHotkeyEvent();
+
+} // namespace FeatureRuntime
